moved assetsData to staticFiles
